+++ Realization/backend/core/core-impl/src/test/java/eu/bcvsolutions/idm/core/model/service/impl/IdentityFilterTest.java
@@ -7,18 +7,21 @@ import java.util.Collections;
 import java.util.UUID;
 
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.transaction.annotation.Transactional;
 
-import eu.bcvsolutions.idm.InitTestData;
+import eu.bcvsolutions.idm.core.api.domain.IdentityState;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityContractDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmTreeNodeDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmTreeTypeDto;
 import eu.bcvsolutions.idm.core.api.dto.filter.IdmIdentityFilter;
+import eu.bcvsolutions.idm.core.api.service.IdmIdentityContractService;
 import eu.bcvsolutions.idm.core.api.service.IdmIdentityService;
 import eu.bcvsolutions.idm.core.model.entity.IdmIdentity_;
 import eu.bcvsolutions.idm.test.api.AbstractIntegrationTest;
@@ -28,21 +31,24 @@ import eu.bcvsolutions.idm.test.api.TestHelper;
  * Test for Identity filters
  *
  * @author Marek Klement
+ * @author Radek Tomiška
  *
  */
+@Transactional
 public class IdentityFilterTest extends AbstractIntegrationTest{
 
-	@Autowired private IdmIdentityService idmIdentityService;
-	@Autowired private TestHelper testHelper;
-
+	@Autowired private IdmIdentityService identityService;
+	@Autowired private IdmIdentityContractService identityContractService;
+	@Autowired private TestHelper helper;
+	
 	@Before
-	public void login(){
-		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
+	public void init() {
+		loginAsAdmin("admin");
 	}
-
+	
 	@After
-	public void logout(){
-		super.logout();
+	public void deleteIdentity() {
+		logout();
 	}
 
 	@Test
@@ -56,7 +62,7 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName001","ThisIsTestName001","ThisIsTestName001@gemail.eu", "000000001", false);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setUsername(person.getUsername());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Username",1, result.getTotalElements());
 		assertEquals("Wrong Username user ID",person.getId(),result.getContent().get(0).getId());
 	}
@@ -66,7 +72,7 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName002","ThisIsTestName002","ThisIsTestName002@gemail.eu", "000000002", false);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setFirstName(person.getFirstName());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Firstname",1, result.getTotalElements());
 		assertEquals("Wrong Firstname user ID",person.getId(),result.getContent().get(0).getId());
 	}
@@ -76,7 +82,7 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName003","ThisIsTestName003","ThisIsTestName003@gemail.eu", "000000003", false);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setLastName(person.getLastName());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Lastname",1, result.getTotalElements());
 		assertEquals("Wrong Lastname user ID",person.getId(),result.getContent().get(0).getId());
 	}
@@ -86,11 +92,11 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName004","ThisIsTestName004","ThisIsTestName004@gemail.eu", "000000004", false);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setDisabled(false);
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
-		person.setDisabled(true);
-		idmIdentityService.save(person);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
+		person.setState(IdentityState.DISABLED);
+		identityService.save(person);
 		filter.setDisabled(false);
-		Page<IdmIdentityDto> result2 = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result2 = identityService.find(filter, null);
 		int changed = (int) (result.getTotalElements() - result2.getTotalElements());
 		assertEquals("Wrong Disabled",1, changed);
 	}
@@ -101,25 +107,25 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setProperty(IdmIdentity_.username.getName());
 		filter.setValue(person.getUsername());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Set Property and Value - username",1, result.getTotalElements());
 		assertEquals("Wrong Property an Value - username user ID",person.getId(),result.getContent().get(0).getId());
 
 		filter.setProperty(IdmIdentity_.firstName.getName());
 		filter.setValue(person.getFirstName());
-		result = idmIdentityService.find(filter, null);
+		result = identityService.find(filter, null);
 		assertEquals("Wrong Set Property and Value - firstname",1, result.getTotalElements());
 		assertEquals("Wrong Property an Value - firstname user ID",person.getId(),result.getContent().get(0).getId());
 
 		filter.setProperty(IdmIdentity_.lastName.getName());
 		filter.setValue(person.getLastName());
-		result = idmIdentityService.find(filter, null);
+		result = identityService.find(filter, null);
 		assertEquals("Wrong Set Property and Value - lastname",1, result.getTotalElements());
 		assertEquals("Wrong Property an Value - lastname user ID",person.getId(),result.getContent().get(0).getId());
 
 		filter.setProperty(IdmIdentity_.email.getName());
 		filter.setValue(person.getEmail());
-		result = idmIdentityService.find(filter, null);
+		result = identityService.find(filter, null);
 		assertEquals("Wrong Set Property and Value - email",1, result.getTotalElements());
 		assertEquals("Wrong Property an Value - email user ID",person.getId(),result.getContent().get(0).getId());
 
@@ -128,15 +134,15 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 	@Test
 	public void testRoleFilter(){
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName013","ThisIsTestName013","ThisIsTestName013@gemail.eu", "100000013", false);
-		IdmTreeTypeDto type = testHelper.createTreeType("ThisIsTestType007");
-		IdmTreeNodeDto node = testHelper.createTreeNode(type,"ThisIsTestNode008",null);
-		IdmIdentityContractDto contract = testHelper.createIdentityContact(person,node);
-		IdmRoleDto somerole = testHelper.createRole();
-		testHelper.createIdentityRole(contract,somerole);
+		IdmTreeTypeDto type = helper.createTreeType("ThisIsTestType007");
+		IdmTreeNodeDto node = helper.createTreeNode(type,"ThisIsTestNode008",null);
+		IdmIdentityContractDto contract = helper.createIdentityContact(person,node);
+		IdmRoleDto somerole = helper.createRole();
+		helper.createIdentityRole(contract,somerole);
 		UUID roleId = somerole.getId();
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setRoles(Collections.singletonList(roleId));
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Roles",1, result.getTotalElements());
 		assertEquals("Wrong Roles user ID",person.getId(),result.getContent().get(0).getId());
 	}
@@ -146,17 +152,17 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName005","ThisIsTestName005","ThisIsTestName005@gemail.eu", "000000005", false);
 		IdmIdentityDto manager = getIdmIdentity("ThisIsTestName006","ThisIsTestName006","ThisIsTestName006@gemail.eu", "000000006", false);
 		UUID manager_id = manager.getId();
-		IdmTreeTypeDto type1 = testHelper.createTreeType("ThisIsTestType001");
-		IdmTreeNodeDto node2 = testHelper.createTreeNode(type1,"Somename002",null);
-		IdmTreeNodeDto node1 = testHelper.createTreeNode(type1,"ThisIsTestNode001",node2);
-		testHelper.createIdentityContact(manager,node2);
-		testHelper.createIdentityContact(person,node1);
+		IdmTreeTypeDto type1 = helper.createTreeType("ThisIsTestType001");
+		IdmTreeNodeDto node2 = helper.createTreeNode(type1,"Somename002",null);
+		IdmTreeNodeDto node1 = helper.createTreeNode(type1,"ThisIsTestNode001",node2);
+		helper.createIdentityContact(manager,node2);
+		helper.createIdentityContact(person,node1);
 		//contract.setGuarantee(manager_id);
 		//IdmIdentityContractDto contract2 = idmIdentityContractService.save(contract);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setSubordinatesFor(manager_id);
 		filter.setSubordinatesByTreeType(type1.getId());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong SubordinatesFor",1, result.getTotalElements());
 		assertEquals("Wrong SuborganizedFor user ID",person.getId(),result.getContent().get(0).getId());
 
@@ -168,18 +174,18 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto manager = getIdmIdentity("ThisIsTestName010","ThisIsTestName010","ThisIsTestName010@gemail.eu", "000000010", false);
 		UUID person_id = person.getId();
 		manager.getId();
-		IdmTreeTypeDto type1 = testHelper.createTreeType("ThisIsTestType004");
-		IdmTreeNodeDto node2 = testHelper.createTreeNode(type1,"Somename001",null);
-		IdmTreeNodeDto node1 = testHelper.createTreeNode(type1,"ThisIsTestNode004",node2);
-		testHelper.createIdentityContact(manager,node2);
-		testHelper.createIdentityContact(person,node1);
+		IdmTreeTypeDto type1 = helper.createTreeType("ThisIsTestType004");
+		IdmTreeNodeDto node2 = helper.createTreeNode(type1,"Somename001",null);
+		IdmTreeNodeDto node1 = helper.createTreeNode(type1,"ThisIsTestNode004",node2);
+		helper.createIdentityContact(manager,node2);
+		helper.createIdentityContact(person,node1);
 		//contract.setGuarantee(manager_id);
 		//IdmIdentityContractDto contract2 = idmIdentityContractService.save(contract);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		//filter.setIncludeGuarantees(true);
 		filter.setManagersFor(person_id);
 		filter.setManagersByTreeType(type1.getId());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong ManagersFor",1, result.getTotalElements());
 		assertEquals("Wrong ManagersFor user ID",manager.getId(),result.getContent().get(0).getId());
 
@@ -190,19 +196,19 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person = getIdmIdentity("aThisIsTestName009","ThisIsTestName009","ThisIsTestName009@gemail.eu", "000000009", false);
 		IdmIdentityDto manager2 = getIdmIdentity("manager009","ThisIsTestName009","ThisIsTestName009@gemail.eu", "000000009", false);
 		IdmIdentityDto manager = getIdmIdentity("ThisIsTestName010","ThisIsTestName010","ThisIsTestName010@gemail.eu", "000000010", false);
-		IdmTreeTypeDto type1 = testHelper.createTreeType("ThisIsTestType0004");
-		IdmTreeNodeDto node1 = testHelper.createTreeNode(type1,"ThisIsTestNode0004",null);
-		IdmTreeNodeDto node2 = testHelper.createTreeNode(type1,"NextThisIsTestNode0004",node1);
-		IdmIdentityContractDto contract = testHelper.createIdentityContact(person,node2);
-		testHelper.createContractGuarantee(contract.getId(),manager2.getId());
-		testHelper.createIdentityContact(manager,node1);
+		IdmTreeTypeDto type1 = helper.createTreeType("ThisIsTestType0004");
+		IdmTreeNodeDto node1 = helper.createTreeNode(type1,"ThisIsTestNode0004",null);
+		IdmTreeNodeDto node2 = helper.createTreeNode(type1,"NextThisIsTestNode0004",node1);
+		IdmIdentityContractDto contract = helper.createIdentityContact(person,node2);
+		helper.createContractGuarantee(contract.getId(),manager2.getId());
+		helper.createIdentityContact(manager,node1);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setManagersFor(person.getId());
 		filter.setIncludeGuarantees(true);
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Managers2For",2, result.getTotalElements());
 		filter.setIncludeGuarantees(false);
-		result = idmIdentityService.find(filter, null);
+		result = identityService.find(filter, null);
 		assertEquals("Wrong Managers2For test 2",1, result.getTotalElements());
 	}
 
@@ -214,8 +220,8 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person3 = getIdmIdentity("Klement","Marek","ThisIsTestName015@gemail.eu", "000000015", false);
 		IdmIdentityDto person4 = getIdmIdentity("Klement","Marek","ThisIsTestName016@gemail.eu", "000000016", false);
 		// setting structure
-		IdmTreeTypeDto type1 = testHelper.createTreeType("ThisIsTestType005x");
-		IdmTreeTypeDto type2 = testHelper.createTreeType("ThisIsTestType006x");
+		IdmTreeTypeDto type1 = helper.createTreeType("ThisIsTestType005x");
+		IdmTreeTypeDto type2 = helper.createTreeType("ThisIsTestType006x");
 		/*
 		      r1  o
 		         /
@@ -223,18 +229,18 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		       /
 	       n2 o
 		 */
-		IdmTreeNodeDto node1 = testHelper.createTreeNode(type1,"ThisIsTestNode005",null);
-		IdmTreeNodeDto node11 = testHelper.createTreeNode(type1,"ThisIsTestNode006",node1);
-		IdmTreeNodeDto node12 = testHelper.createTreeNode(type2,"ThisIsTestNode007",node11);
+		IdmTreeNodeDto node1 = helper.createTreeNode(type1,"ThisIsTestNode005",null);
+		IdmTreeNodeDto node11 = helper.createTreeNode(type1,"ThisIsTestNode006",node1);
+		IdmTreeNodeDto node12 = helper.createTreeNode(type2,"ThisIsTestNode007",node11);
 		/*
 		    r2  o
 		 */
-		IdmTreeNodeDto node2 = testHelper.createTreeNode(type2,"ThisIsTestNode008",null);
+		IdmTreeNodeDto node2 = helper.createTreeNode(type2,"ThisIsTestNode008",null);
 		// contracts
-		testHelper.createIdentityContact(person1,node1);
-		testHelper.createIdentityContact(person2,node11);
-		testHelper.createIdentityContact(person3,node12);
-		testHelper.createIdentityContact(person4,node2);
+		helper.createIdentityContact(person1,node1);
+		helper.createIdentityContact(person2,node11);
+		helper.createIdentityContact(person3,node12);
+		helper.createIdentityContact(person4,node2);
 		// node1 UUID
 		UUID node1id = node1.getId();
 		// test
@@ -242,37 +248,37 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		filter.setFirstName(person1.getFirstName());
 		filter.setRecursively(true);
 		filter.setTreeNode(node1id);
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong Recursive firstname",3, result.getTotalElements());
 		filter.setRecursively(false);
-		result = idmIdentityService.find(filter, null);
+		result = identityService.find(filter, null);
 		assertEquals("Wrong NonRecursive firstname",1, result.getTotalElements());
 	}
 
 	@Test
 	public void testTreeTypeFilter(){
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName007","ThisIsTestName007","ThisIsTestName007@gemail.eu", "000000007", false);
-		IdmTreeTypeDto type = testHelper.createTreeType("ThisIsTestType002");
-		IdmTreeNodeDto node = testHelper.createTreeNode(type,"ThisIsTestNode002",null);
+		IdmTreeTypeDto type = helper.createTreeType("ThisIsTestType002");
+		IdmTreeNodeDto node = helper.createTreeNode(type,"ThisIsTestNode002",null);
 		UUID typeUuid = type.getId();
-		testHelper.createIdentityContact(person,node);
+		helper.createIdentityContact(person,node);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setTreeType(typeUuid);
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong TreeType",1, result.getTotalElements());
 	}
 
 	@Test
 	public void testTreeNodeFilter(){
 		IdmIdentityDto person = getIdmIdentity("ThisIsTestName008","ThisIsTestName008","ThisIsTestName008@gemail.eu", "000000008", false);
-		IdmTreeTypeDto type = testHelper.createTreeType("ThisIsTestType003");
-		IdmTreeNodeDto node = testHelper.createTreeNode(type,"ThisIsTestNode003",null);
+		IdmTreeTypeDto type = helper.createTreeType("ThisIsTestType003");
+		IdmTreeNodeDto node = helper.createTreeNode(type,"ThisIsTestNode003",null);
 		UUID nodeUuid = node.getId();
-		testHelper.createIdentityContact(person,node);
+		helper.createIdentityContact(person,node);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setTreeNode(nodeUuid);
 		filter.setRecursively(false);
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter, null);
+		Page<IdmIdentityDto> result = identityService.find(filter, null);
 		assertEquals("Wrong TreeNode",1, result.getTotalElements());
 	}
 
@@ -280,7 +286,7 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 	public void testNothingToFind(){
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setFirstName("Adolf");
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter,null);
+		Page<IdmIdentityDto> result = identityService.find(filter,null);
 		assertEquals("Wrong blank filter",0, result.getTotalElements());
 	}
 
@@ -293,7 +299,7 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setFirstName(person1.getFirstName());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter,null);
+		Page<IdmIdentityDto> result = identityService.find(filter,null);
 		assertEquals("Wrong more to find filter",3, result.getTotalElements());
 	}
 
@@ -307,10 +313,10 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		filter.setFirstName(person1.getFirstName());
 		filter.setLastName(person2.getLastName());
 		filter.setDisabled(false);
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter,null);
+		Page<IdmIdentityDto> result = identityService.find(filter,null);
 		assertEquals("Wrong more filter",3, result.getTotalElements());
 		filter.setUsername(person3.getUsername());
-		result = idmIdentityService.find(filter,null);
+		result = identityService.find(filter,null);
 		assertEquals("Wrong more filter",1, result.getTotalElements());
 		assertEquals("Wrong more filter user ID",person3.getId(),result.getContent().get(0).getId());
 	}
@@ -319,18 +325,18 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 	public void testSameButDifferentTreeTypeFilter(){
 		IdmIdentityDto person1 = getIdmIdentity("thistn01","thistn01","thistn02@a.eu", "000010001", false);
 		getIdmIdentity("thistn01","thistn01","thistn02@a.eu", "000010001", false);
-		IdmTreeTypeDto type1 = testHelper.createTreeType("typeName001");
-		IdmTreeTypeDto type2 = testHelper.createTreeType("typeName002");
-		IdmTreeNodeDto	 node1 = testHelper.createTreeNode(type1,"nodeName001",null);
-		testHelper.createTreeNode(type2,"nodeName002",null);
-		testHelper.createIdentityContact(person1,node1);
-		testHelper.createIdentityContact(person1,node1);
+		IdmTreeTypeDto type1 = helper.createTreeType("typeName001");
+		IdmTreeTypeDto type2 = helper.createTreeType("typeName002");
+		IdmTreeNodeDto	 node1 = helper.createTreeNode(type1,"nodeName001",null);
+		helper.createTreeNode(type2,"nodeName002",null);
+		helper.createIdentityContact(person1,node1);
+		helper.createIdentityContact(person1,node1);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setFirstName(person1.getFirstName());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter,null);
+		Page<IdmIdentityDto> result = identityService.find(filter,null);
 		assertEquals("Wrong sameButDiff filter",2, result.getTotalElements());
 		filter.setTreeType(type1.getId());
-		result = idmIdentityService.find(filter,null);
+		result = identityService.find(filter,null);
 		assertEquals("Wrong sameButDiff filter 2",1, result.getTotalElements());
 	}
 
@@ -342,32 +348,63 @@ public class IdentityFilterTest extends AbstractIntegrationTest{
 		IdmIdentityDto person4 = getIdmIdentity("a2testuser","b3testuser","b4testuser@a.eu", "333010005", false);
 		IdmIdentityFilter filter = new IdmIdentityFilter();
 		filter.setText(person1.getFirstName());
-		Page<IdmIdentityDto> result = idmIdentityService.find(filter,null);
+		Page<IdmIdentityDto> result = identityService.find(filter,null);
 		assertEquals("Wrong Text filter - firstname",3, result.getTotalElements());
 		filter.setText(person2.getLastName());
-		result = idmIdentityService.find(filter,null);
+		result = identityService.find(filter,null);
 		assertEquals("Wrong Text filter - lastname",2, result.getTotalElements());
 		filter.setText(person3.getEmail());
-		result = idmIdentityService.find(filter,null);
+		result = identityService.find(filter,null);
 		assertEquals("Wrong Text filter - email",1, result.getTotalElements());
 		filter.setText(person4.getUsername());
-		result = idmIdentityService.find(filter,null);
+		result = identityService.find(filter,null);
 		assertEquals("Wrong Text filter - username",person4.getId(), result.getContent().get(0).getId());
 		person1.setDescription("This is describtion");
-		IdmIdentityDto person5 = idmIdentityService.save(person1);
+		IdmIdentityDto person5 = identityService.save(person1);
 		filter.setText(person5.getDescription());
-		result = idmIdentityService.find(filter,null);
+		result = identityService.find(filter,null);
 		assertEquals("Wrong Text filter - describtion",person1.getId(), result.getContent().get(0).getId());
+	}
+	
+	@Test
+	public void testStateFilter() {
+		String firstname = helper.createName();
+		IdmIdentityDto identityOne = new IdmIdentityDto();
+		identityOne.setUsername(helper.createName());
+		identityOne.setFirstName(firstname);
+		identityOne = identityService.save(identityOne);
+		identityContractService.delete(helper.getPrimeContract(identityOne.getId()));
+		IdmIdentityDto identity = new IdmIdentityDto();
+		identity.setUsername(helper.createName());
+		identity.setFirstName(firstname);
+		identity.setState(IdentityState.DISABLED_MANUALLY);
+		identityService.save(identity);
+		//
+		IdmIdentityFilter filter = new IdmIdentityFilter();
+		filter.setFirstName(firstname);
+		filter.setState(IdentityState.VALID);
+		Page<IdmIdentityDto> results = identityService.find(filter, null);
+		Assert.assertTrue(results.getTotalElements() == 0);
+		//
+		filter.setState(IdentityState.NO_CONTRACT);
+		results = identityService.find(filter, null);
+		Assert.assertTrue(results.getTotalElements() == 1);
+		Assert.assertEquals(identityOne.getUsername(), results.getContent().get(0).getUsername());
+		//
+		filter.setState(IdentityState.DISABLED_MANUALLY);
+		results = identityService.find(filter, null);
+		Assert.assertTrue(results.getTotalElements() == 1);
+		Assert.assertEquals(identity.getUsername(), results.getContent().get(0).getUsername());
 	}
 
 	private IdmIdentityDto getIdmIdentity(String firstName, String lastName, String email, String phone, boolean disabled){
-		IdmIdentityDto identity2 = testHelper.createIdentity();
+		IdmIdentityDto identity2 = helper.createIdentity();
 		identity2.setFirstName(firstName);
 		identity2.setLastName(lastName);
 		identity2.setEmail(email);
-		identity2.setDisabled(disabled);
+		identity2.setState(disabled ? IdentityState.DISABLED : IdentityState.VALID);
 		identity2.setPhone(phone);
-		return idmIdentityService.save(identity2);
+		return identityService.save(identity2);
 	}
 
 }